+++ scripts/build-sdist.py
@@ -7,8 +7,11 @@ PROJECTION_PATH = (Path(__file__).parent.parent / "projection").resolve()
 
 try:
     import winrt_sdk  # noqa: F401
+    import winrt_windows_app_sdk  # noqa: F401
 except ImportError:
-    print("winrt_sdk must be installed first installed in order for build to succeed.")
+    print(
+        "winrt_sdk and winrt_windows_app_sdk must be installed first installed in order for build to succeed."
+    )
     exit(1)
 
 
@@ -16,6 +19,7 @@ for package_path in chain(
     [str(PROJECTION_PATH / "winrt-runtime")],
     iglob(str(PROJECTION_PATH / "interop" / "winrt-Windows*")),
     iglob(str(PROJECTION_PATH / "winrt-Windows*")),
+    iglob(str(PROJECTION_PATH / "winrt-Microsoft*")),
 ):
     subprocess.check_call(
         [